requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.16.0

# PDF Processing & OCR
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# AI / NLP
anthropic==0.39.0
openai==1.6.1
spacy==3.7.2
transformers==4.36.2
torch==2.1.2

# Fuzzy Matching & Data Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
phonenumbers==8.13.27
usaddress==0.5.10
nameparser==1.1.3
jellyfish==1.0.3

# Data Science & ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Geocoding & Mapping
geopy==2.4.1
folium==0.15.1

# Graph Processing (for family trees)
networkx==3.2.1

# API Integrations
twilio==8.11.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# Caching
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.0

# Validation
cerberus==1.3.5
validators==0.22.0

# Logging
colorlog==6.8.0

# Excel Support
openpyxl==3.1.2
xlrd==2.0.1

# Data Export
reportlab==4.0.7
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
